Update ez-database to use actively maintained libraries
- Move from clojure.java.jdbc to next.jdbc
- Move from honeysql v1 to honeysql v2
- Move from yesql to HugSQL

clojure.java.jdbc is considered Stable, and no longer actively
maintained. next.jdbc is the evolution of that library, giving a
number of bells and whistles to use compared to clojure.java.jdbc.

HoneySQL moved from v1 to v2, with a couple of breaking changes, but
with a greater deal of flexibility.

With the move to next.jdbc, yesql no longer works, as it's not been
actively updated. HugSQL offers a rough equalivent library that fills
the same need.